feat: MethodAttributesSupport trait introduced and moved related functions and properties
